move tokenization helper functions to utils; streamline KL calc for KTO
