+++ tests/test_dpo_trainer.py
@@ -31,7 +31,7 @@ from transformers import (
 from transformers.testing_utils import require_bitsandbytes, require_peft
 
 from trl import DPOConfig, DPOTrainer, FDivergenceType
-from trl.trainer.dpo_trainer import _build_tokenized_answer, _truncate_tokens
+from trl.trainer.utils import _build_tokenized_answer, _truncate_tokens
 
 from .testing_utils import require_no_wandb
