+++ trl/trainer/dpo_trainer.py
@@ -62,6 +62,13 @@ from .utils import (
     generate_model_card,
     pad_to_length,
     peft_module_casting_to_bf16,
+    _add_special_tokens,
+    _adjust_prompt_length,
+    _append_prompt_tokens_to_batch,
+    _build_sequence_tokens,
+    _process_answer,
+    _process_prompt,
+    _truncate_tokens,
 )
 
 
@@ -71,6 +78,7 @@ if is_peft_available():
 
 if is_wandb_available():
     import wandb
+
 
 if is_deepspeed_available():
     import deepspeed
@@ -117,141 +125,6 @@ def _tokenize(
     return dict(batch)
 
 
-def _process_prompt(
-    prompts: List[str], processor: Optional[Callable], tokenizer: PreTrainedTokenizerBase, images: List[Optional[Any]]
-) -> List[Dict[str, List[int]]]:
-    """
-    Processes a list of prompts by tokenizing them, optionally using a processor for additional processing.
-    """
-    if processor:
-        processor_kwargs = (
-            {"add_special_tokens": False} if "add_special_tokens" in inspect.signature(processor).parameters else {}
-        )
-        prompt_tokens = []
-        for prompt, image in zip(prompts, images):
-            tokens = processor(images=image, text=prompt, **processor_kwargs)
-            tokens = {k: v[0] for k, v in tokens.items()}
-            if not isinstance(tokens["input_ids"], list):
-                tokens["input_ids"] = tokens["input_ids"].tolist()
-                tokens["attention_mask"] = tokens["attention_mask"].tolist()
-            prompt_tokens.append(tokens)
-    else:
-        prompt_tokens = [tokenizer(prompt, add_special_tokens=False) for prompt in prompts]
-    return [{f"prompt_{k}": v for k, v in tokens.items()} for tokens in prompt_tokens]
-
-
-def _process_answer(
-    prompts: List[str],
-    answers: List[str],
-    processor: Optional[Callable],
-    tokenizer: PreTrainedTokenizerBase,
-    images: List[Optional[Any]],
-) -> List[Dict[str, Any]]:
-    return [
-        _build_tokenized_answer(prompt, answer, image, processor=processor, tokenizer=tokenizer)
-        for prompt, answer, image in zip(prompts, answers, images)
-    ]
-
-
-def _adjust_prompt_length(
-    prompt_tokens: List[Dict[str, List[int]]],
-    chosen_tokens: List[Dict[str, List[int]]],
-    rejected_tokens: List[Dict[str, List[int]]],
-) -> List[int]:
-    prompt_len_input_ids = []
-    for p_tokens, c_tokens, r_tokens in zip(prompt_tokens, chosen_tokens, rejected_tokens):
-        c_len = len(c_tokens["prompt_input_ids"])
-        r_len = len(r_tokens["prompt_input_ids"])
-        min_len = min(c_len, r_len)
-
-        for k, v in p_tokens.items():
-            p_tokens[k] = v[:min_len]
-
-        num_diff_tokens = sum([a != b for a, b in zip(c_tokens["prompt_input_ids"], r_tokens["prompt_input_ids"])])
-        num_diff_len = abs(c_len - r_len)
-        if num_diff_tokens > 1 or num_diff_len > 1:
-            raise ValueError(
-                "Chosen and rejected prompt_input_ids might only differ on the last token due to tokenizer merge ops."
-            )
-        prompt_len_input_ids.append(min_len)
-    return prompt_len_input_ids
-
-
-def _add_special_tokens(
-    tokenizer: PreTrainedTokenizerBase,
-    prompt_len_input_ids: List[int],
-    prompt_tokens: List[Dict[str, List[int]]],
-    chosen_tokens: List[Dict[str, List[int]]],
-    rejected_tokens: List[Dict[str, List[int]]],
-) -> Tuple[List[Dict[str, List[int]]], List[Dict[str, List[int]]], List[Dict[str, List[int]]]]:
-    for i in range(len(prompt_tokens)):
-        prompt_tokens[i], chosen_tokens[i], rejected_tokens[i] = add_bos_token_if_needed(
-            tokenizer.bos_token_id,
-            prompt_len_input_ids[i],
-            prompt_tokens[i],
-            len(chosen_tokens[i]["prompt_input_ids"]),
-            chosen_tokens[i],
-            len(rejected_tokens[i]["prompt_input_ids"]),
-            rejected_tokens[i],
-        )
-
-        chosen_tokens[i], rejected_tokens[i] = add_eos_token_if_needed(
-            tokenizer.eos_token_id, chosen_tokens[i], rejected_tokens[i]
-        )
-    return prompt_tokens, chosen_tokens, rejected_tokens
-
-
-def _truncate_tokens(
-    chosen_tokens: List[Dict[str, List[int]]],
-    rejected_tokens: List[Dict[str, List[int]]],
-    prompt_tokens: List[Dict[str, List[int]]],
-    args: DPOConfig,
-) -> None:
-    """
-    Truncates the tokens in chosen, rejected, and prompt sequences to ensure they fit within the maximum length constraints.
-    """
-    if args.truncation_mode not in ["keep_start", "keep_end"]:
-        raise ValueError(f"Invalid truncation mode: {args.truncation_mode}")
-
-    for c_tokens, r_tokens, p_tokens in zip(chosen_tokens, rejected_tokens, prompt_tokens):
-        longer_response_length = max(len(c_tokens["input_ids"]), len(r_tokens["input_ids"]))
-
-        # if combined sequence is too long, truncate the prompt
-        for answer_tokens in [c_tokens, r_tokens, p_tokens]:
-            if len(answer_tokens["prompt_input_ids"]) + longer_response_length > args.max_length:
-                if args.truncation_mode == "keep_start":
-                    for k in ["prompt_input_ids", "prompt_attention_mask"]:
-                        answer_tokens[k] = answer_tokens[k][: args.max_prompt_length]
-                elif args.truncation_mode == "keep_end":
-                    for k in ["prompt_input_ids", "prompt_attention_mask"]:
-                        answer_tokens[k] = answer_tokens[k][-args.max_prompt_length :]
-
-        # if that's still too long, truncate the response from the end
-        for answer_tokens in [c_tokens, r_tokens]:
-            if len(answer_tokens["prompt_input_ids"]) + longer_response_length > args.max_length:
-                for k in ["input_ids", "attention_mask"]:
-                    answer_tokens[k] = answer_tokens[k][: args.max_length - args.max_prompt_length]
-
-
-def _build_sequence_tokens(
-    batch: Dict[str, List[int]], tokens: List[Dict[str, List[int]]], args: DPOConfig, prefix: str
-) -> None:
-    for token in tokens:
-        sequence_tokens = {f"{prefix}_{k}": token[f"prompt_{k}"] + token[k] for k in ["input_ids", "attention_mask"]}
-        sequence_tokens[f"{prefix}_labels"] = sequence_tokens[f"{prefix}_input_ids"][:]
-        sequence_tokens[f"{prefix}_labels"][: len(token["prompt_input_ids"])] = [args.label_pad_token_id] * len(
-            token["prompt_input_ids"]
-        )
-        for k, v in sequence_tokens.items():
-            batch[k].append(v)
-
-
-def _append_prompt_tokens_to_batch(batch: Dict[str, List[int]], prompt_tokens: List[Dict[str, List[int]]]) -> None:
-    for p_tokens in prompt_tokens:
-        for k, v in p_tokens.items():
-            batch[k].append(v)
-
-
 def _tokenize_encoder_decoder(
     batch: Dict[str, List[int]],
     tokenizer: PreTrainedTokenizerBase,
@@ -288,74 +161,6 @@ def _tokenize_encoder_decoder(
         batch["chosen_decoder_input_ids"] = model.prepare_decoder_input_ids_from_labels(
             labels=torch.tensor(batch["chosen_labels"])
         )
-
-
-def _build_tokenized_answer(
-    prompt: str,
-    answer: str,
-    images: Optional[List[Any]] = None,
-    processor: Optional[Callable] = None,
-    tokenizer: Optional[PreTrainedTokenizerBase] = None,
-) -> Dict[str, Any]:
-    """
-    Build tokenized response, handling vision models and different tokenizers.
-    """
-
-    def tokenize(text, images=None):
-        if processor:
-            processor_kwargs = (
-                {"add_special_tokens": False}
-                if "add_special_tokens" in inspect.signature(processor).parameters
-                else {}
-            )
-            tokenized = processor(images=images, text=text, **processor_kwargs)
-            tokenized = {k: v[0] for k, v in tokenized.items()}
-            if not isinstance(tokenized["input_ids"], list):
-                tokenized["input_ids"] = tokenized["input_ids"].tolist()
-                tokenized["attention_mask"] = tokenized["attention_mask"].tolist()
-        else:
-            tokenized = tokenizer(text, add_special_tokens=False)
-        return tokenized
-
-    full_tokenized = tokenize(prompt + answer, images)
-    prompt_tokenized = tokenize(prompt, images)
-
-    prompt_input_ids = prompt_tokenized["input_ids"]
-    answer_input_ids = full_tokenized["input_ids"][len(prompt_input_ids) :]
-    answer_attention_mask = full_tokenized["attention_mask"][len(prompt_input_ids) :]
-
-    if len(full_tokenized["input_ids"]) != len(prompt_input_ids + answer_input_ids):
-        raise ValueError("Prompt input ids and answer input ids should have the same length.")
-
-    # On some tokenizers, like Llama-2 tokenizer, there are occasions where tokens
-    # can be merged together when tokenizing prompt+answer. This could result
-    # on the last token from the prompt being different when tokenized on its own
-    # vs when done as prompt+answer.
-    response_token_ids_start_idx = len(prompt_input_ids)
-
-    # If tokenized prompt is different than both prompt+answer, then it means the
-    # last token has changed due to merging.
-    if prompt_input_ids != full_tokenized["input_ids"][:response_token_ids_start_idx]:
-        response_token_ids_start_idx -= 1
-
-    prompt_input_ids = full_tokenized["input_ids"][:response_token_ids_start_idx]
-    prompt_attention_mask = full_tokenized["attention_mask"][:response_token_ids_start_idx]
-
-    if len(prompt_input_ids) != len(prompt_attention_mask):
-        raise ValueError("Prompt input ids and attention mask should have the same length.")
-
-    return_dict = {
-        "prompt_input_ids": prompt_input_ids,
-        "prompt_attention_mask": prompt_attention_mask,
-        "input_ids": answer_input_ids,
-        "attention_mask": answer_attention_mask,
-    }
-    if "pixel_values" in full_tokenized:
-        return_dict["prompt_pixel_values"] = full_tokenized["pixel_values"]
-    if "pixel_attention_mask" in full_tokenized:
-        return_dict["prompt_pixel_attention_mask"] = full_tokenized["pixel_attention_mask"]
-
-    return return_dict
 
 
 class DPOTrainer(Trainer):
